+++ plom/create/status.py
@@ -1,9 +1,12 @@
 # SPDX-License-Identifier: AGPL-3.0-or-later
 # Copyright (C) 2022 Colin B. Macdonald
 
+from textwrap import wrap
+
 from plom import SpecVerifier
 from plom.create import with_manager_messenger
 from plom.plom_exceptions import PlomServerNotReady, PlomBenignException
+from plom.misc_utils import format_int_list_with_runs
 
 
 # support for colour checkmarks
@@ -88,6 +91,32 @@ def status(*, msgr):
     except PlomBenignException as e:
         # TODO: make PlomServerHasNoClasslist: subclass of PlomServerNotReady?
         print(cross + f" No classlist: {e}")
+        classlist = None
+    if classlist and classlist[0].get("paper_number", None):
+        # classlist has papernum column
+        papernums = [
+            int(r["paper_number"]) for r in classlist if int(r["paper_number"]) > 0
+        ]
+        if len(set(papernums)) != len(papernums):
+            print(cross + ' "paper_number" fields are not unique!')
+        else:
+            print(
+                check_mark
+                + f' {len(papernums)} have "paper_number" field for prenaming:'
+            )
+            print(
+                "    " + "\n    ".join(wrap(format_int_list_with_runs(papernums), 72))
+            )
+    else:
+        papernums = []
+    if spec:
+        not_named = set(range(1, spec["numberToProduce"]))
+        not_named.difference_update(papernums)
+        print(
+            check_mark
+            + f" {len(not_named)} unassigned paper numbers (these will not be prenamed):"
+        )
+        print("    " + "\n    ".join(wrap(format_int_list_with_runs(not_named), 72)))
 
     print("\nDatabase")
     print("---------\n")