+++ plom/specVerifier.py
@@ -160,7 +160,6 @@ class SpecVerifier:
     ... 'numberOfQuestions': 3,
     ... 'totalMarks': 25,
     ... 'numberToProduce': 20,
-    ... 'numberToName': 10,
     ... 'privateSeed': '1001378822317872',
     ... 'publicCode': '270385',
     ... 'idPage': 1,
@@ -180,7 +179,6 @@ class SpecVerifier:
       Long name of exam = Midterm Demo using Plom
       Number of source versions = 2
       Number of tests to produce = 20
-      Number of those to be printed with names = 10
       Number of pages = 6
       IDpage = 1
       Do not mark pages = [2]
@@ -271,7 +269,7 @@ class SpecVerifier:
         if num_to_produce:
             from plom.create.demotools import getDemoClassListLength
 
-            # TODO: 20 and 10 in source file hardcoded here, use regex instead
+            # TODO: 20 in source file hardcoded here, use regex instead
             s = s.replace(
                 "numberToProduce = 20",
                 "numberToProduce = {}".format(num_to_produce),
@@ -281,10 +279,6 @@ class SpecVerifier:
                 raise ValueError(
                     "Demo size capped at classlist length of {}".format(classlist_len)
                 )
-            s = s.replace(
-                "numberToName = 10",
-                "numberToName = {}".format(min(num_to_produce // 2, classlist_len)),
-            )
         return s
 
     @classmethod
@@ -310,24 +304,8 @@ class SpecVerifier:
     def number_to_produce(self):
         return self.spec["numberToProduce"]
 
-    @property
-    def number_to_name(self):
-        return self.spec["numberToName"]
-
     # aliases to match the toml file
     numberToProduce = number_to_produce
-    numberToName = number_to_name
-
-    def set_number_papers_to_name(self, n):
-        """Set previously-deferred number of named papers.
-
-        exceptions:
-            ValueError: number of named papers already set.
-        """
-        if self.numberToName >= 0:
-            raise ValueError("Number of named papers already set: read-only")
-        self.spec["numberToName"] = n
-        log.info('deferred number of named papers now set to "{}"'.format(n))
 
     def set_number_papers_add_spares(
         self, n, spare_percent=10, min_extra=5, max_extra=100
@@ -369,9 +347,6 @@ class SpecVerifier:
                 # "Public code (to prevent project collisions) = {}".format(self.spec["publicCode"]),
                 # "Private random seed (for randomisation) = {}".format(self.spec["privateSeed"]),
                 "Number of tests to produce = {}".format(self.numberToProduce),
-                "Number of those to be printed with names = {}".format(
-                    self.numberToName
-                ),
                 "Number of pages = {}".format(self.spec["numberOfPages"]),
                 "IDpage = {}".format(self.spec["idPage"]),
                 "Do not mark pages = {}".format(self.spec["doNotMarkPages"]),
@@ -504,7 +479,6 @@ class SpecVerifier:
             "numberOfVersions",
             "numberOfPages",
             "numberToProduce",
-            "numberToName",
             "idPage",
         ]:
             if x not in self.spec:
@@ -519,6 +493,12 @@ class SpecVerifier:
         for x in ["doNotMarkPages", "totalMarks", "numberOfQuestions"]:
             if x in self.spec:
                 print(f'  contains "{x}"{chk}')
+        # check for no longer supported numberToName field
+        if "numberToName" in self.spec:
+            raise NotImplementedError(
+                'The "numberToName" spec-field is removed in favour of'
+                f' the "paper_number" column in the classlist. {self.spec}'
+            )
 
     def check_name_and_production_numbers(self, print=print):
         print("Checking specification name and numbers")
@@ -544,7 +524,7 @@ class SpecVerifier:
                 )
             print('    "{}" = {} is positive integer{}'.format(x, self.spec[x], chk))
 
-        for x in ("numberToName", "numberToProduce"):
+        for x in ["numberToProduce"]:
             try:
                 self.spec[x] = int(self.spec[x])
             except ValueError:
@@ -561,22 +541,6 @@ class SpecVerifier:
 
         if self.numberToProduce == 0:
             raise ValueError('Specification error - "numberToProduce" cannot be zero.')
-
-        if self.numberToProduce > 0:
-            if self.numberToProduce < self.numberToName:
-                raise ValueError(
-                    "Specification error - insufficient papers: producing fewer papers {} than you wish to name {}. Produce more papers.".format(
-                        self.numberToProduce, self.numberToName
-                    )
-                )
-            print("    Producing enough papers to cover named papers" + chk)
-            if self.numberToProduce < 1.05 * self.numberToName:
-                print(
-                    "WARNING: you are producing less than 5% un-named papers; you may want more spares"
-                    + warn_mark
-                )
-            else:
-                print("    Producing sufficiently many spare papers" + chk)
 
     def check_questions(self, print=print):
         if "numberOfQuestions" not in self.spec: